Splitting RecursiveComplexTypesTests into E2E and unit tests
