+++ test/E2ETest/Microsoft.Test.E2E.AspNet.OData/Build.AspNet/Microsoft.Test.E2E.AspNet.OData.csproj
@@ -1411,8 +1411,11 @@
     <Compile Include="..\ModelBuilder\PropertyTests.cs">
       <Link>ModelBuilder\PropertyTests.cs</Link>
     </Compile>
-    <Compile Include="..\ModelBuilder\RecursiveComplexTypesTests.cs">
-      <Link>ModelBuilder\RecursiveComplexTypesTests.cs</Link>
+    <Compile Include="..\..\..\TestModels\RecursiveComplexTypesModels.cs" >
+      <Link>ModelBuilder\RecursiveComplexTypesModels.cs</Link>
+    </Compile>
+    <Compile Include="..\ModelBuilder\RecursiveComplexTypesE2ETests.cs">
+      <Link>ModelBuilder\RecursiveComplexTypesE2ETests.cs</Link>
     </Compile>
     <Compile Include="..\ModelBuilder\SpecialCharactersLinkGenerationTests.cs">
       <Link>ModelBuilder\SpecialCharactersLinkGenerationTests.cs</Link>
