+++ test/UnitTest/Microsoft.AspNet.OData.Test.Shared/Microsoft.AspNet.OData.Test.Shared.projitems
@@ -34,6 +34,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Builder\ComplexTypeConfigurationTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\ComplexTypeTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\ContainmentPathBuilderTest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Builder\Conventions\RecursiveComplexTypesTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\DynamicPropertyDictionaryAnnotationTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\EdmModelAsserts.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\EdmTypeConfigurationExtensionsTest.cs" />
@@ -102,6 +103,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)Builder\TestModels\OrderHeader.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\TestModels\OrderLine.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\TestModels\PropertyAlias.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)\..\..\TestModels\RecursiveComplexTypesModels.cs">
+      <Link>Builder\TestModels\RecursiveComplexTypesModels.cs</Link>
+    </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Builder\TestModels\RecursiveModel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\TestModels\SpecialOrderLine.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Builder\TestModels\Address.cs" />
